misc: dedup sigchld helper, bossbar open_url, mcp test helpers (#3958)
Fold the byte-identical twins flagged by the embedding dupe finder:

- ensure_sigchld_default: moved from plumb/ex and tui/ex into
  unibind-ex-runtime (the BEAM-side rustler runtime crate) as
  unibind_ex_runtime::ensure_sigchld_default; both NIF crates call it and
  drop their local copy plus their now-unused direct libc dep.
- open_url: moved into overlay-core (shared by both overlays) as
  overlay_core::open_url(app, url); folded all three copies (orb,
  bossbar/overlay.rs, bossbar/layer_shell.rs), app tag keeps each log prefix.
- _spool_home: the weave spool teardown fixture now lives once in
  packages/mcp/src/weave/conftest.py, copied into the fabric test derivation
  via a weave passthru; removed from test_weave.py and test_fabric.py.
- test_type_hints_explicit: extracted to packages/mcp/tests/type_hint_support.py
  (mirroring linear_test_support.py) and wired into the notion, iphone, and
  slack test derivations; folded all three copies.

(authored by Claude Code, model Opus)
